Move storage and loading of statis items to separate class for testing.
