+++ YUViewLib/src/statistics/StatisticsType.cpp
@@ -37,95 +37,6 @@
 namespace stats
 {
 
-namespace
-{
-
-// Get a string with all values of the QPen
-QString convertPenToString(const LineDrawStyle &style)
-{
-  auto colorHex  = QString::fromStdString(style.color.toHex());
-  auto patternIt = std::find(AllPatterns.begin(), AllPatterns.end(), style.pattern);
-  if (patternIt == AllPatterns.end())
-    return {};
-  auto patternIdx = std::distance(AllPatterns.begin(), patternIt);
-  return QString("%1 %2 %3").arg(colorHex).arg(style.width).arg(patternIdx);
-}
-
-// The inverse functio to get a QPen from the string
-LineDrawStyle convertStringToPen(const QString &str)
-{
-  LineDrawStyle style;
-  QStringList   split = str.split(" ");
-  if (split.length() == 3)
-  {
-    style.color     = Color(split[0].toStdString());
-    style.width     = split[1].toDouble();
-    auto patternIdx = split[2].toInt();
-    if (patternIdx >= 0 && unsigned(patternIdx) < AllPatterns.size())
-      style.pattern = AllPatterns[patternIdx];
-  }
-  return style;
-}
-
-void addModifiedValuesToElement(YUViewDomElement                                      &element,
-                                const std::optional<StatisticsType::ValueDataOptions> &options,
-                                const std::optional<StatisticsType::ValueDataOptions> &initOptions)
-{
-  if (!options)
-    return;
-
-  if (!initOptions || options->render != initOptions->render)
-    element.setAttribute("renderValueData", options->render);
-  if (!initOptions || options->scaleToBlockSize != initOptions->scaleToBlockSize)
-    element.setAttribute("scaleValueToBlockSize", options->render);
-  if (!initOptions || options->colorMapper != initOptions->colorMapper)
-    options->colorMapper.savePlaylist(element);
-}
-
-std::vector<StatisticsType::ArrowHead> AllArrowHeads = {StatisticsType::ArrowHead::arrow,
-                                                        StatisticsType::ArrowHead::circle,
-                                                        StatisticsType::ArrowHead::none};
-
-void addModifiedValuesToElement(YUViewDomElement                                       &element,
-                                const std::optional<StatisticsType::VectorDataOptions> &options,
-                                const std::optional<StatisticsType::VectorDataOptions> &initOptions)
-{
-  if (!options)
-    return;
-
-  if (!initOptions || options->render != initOptions->render)
-    element.setAttribute("renderVectorData", options->render);
-  if (!initOptions || options->renderDataValues != initOptions->renderDataValues)
-    element.setAttribute("renderVectorDataValues", options->renderDataValues);
-  if (!initOptions || options->scaleToZoom != initOptions->scaleToZoom)
-    element.setAttribute("scaleVectorToZoom", options->scaleToZoom);
-  if (!initOptions || options->style != initOptions->style)
-    element.setAttribute("vectorStyle", convertPenToString(options->style));
-  if (!initOptions || options->scale != initOptions->scale)
-    element.setAttribute("vectorScale", options->scale);
-  if (!initOptions || options->mapToColor != initOptions->mapToColor)
-    element.setAttribute("mapVectorToColor", options->mapToColor);
-  if (!initOptions || options->arrowHead != initOptions->arrowHead)
-  {
-    if (const auto index = vectorIndexOf(AllArrowHeads, options->arrowHead))
-      element.setAttribute("renderarrowHead", static_cast<int>(*index));
-  }
-}
-
-void addModifiedValuesToElement(YUViewDomElement                  &element,
-                                const StatisticsType::GridOptions &options,
-                                const StatisticsType::GridOptions &initOptions)
-{
-  if (options.render != initOptions.render)
-    element.setAttribute("renderGrid", options.render);
-  if (options.style != initOptions.style)
-    element.setAttribute("gridStyle", convertPenToString(options.style));
-  if (options.scaleToZoom != initOptions.scaleToZoom)
-    element.setAttribute("scaleGridToZoom", options.scaleToZoom);
-}
-
-} // namespace
-
 bool LineDrawStyle::operator==(const LineDrawStyle &other) const
 {
   return color == other.color && width == other.width && pattern == other.pattern;
@@ -134,101 +45,6 @@ bool LineDrawStyle::operator==(const LineDrawStyle &other) const
 bool LineDrawStyle::operator!=(const LineDrawStyle &other) const
 {
   return !(*this == other);
-}
-
-/* Save all the settings of the statistics type that have changed from the initial state
- */
-void StatisticsType::saveToPlaylist(YUViewDomElement &root) const
-{
-  bool allValuesIdenticalToInitialValues = (init.render == this->render &&                       //
-                                            init.alphaFactor == this->alphaFactor &&             //
-                                            init.valueDataOptions == this->valueDataOptions &&   //
-                                            init.vectorDataOptions == this->vectorDataOptions && //
-                                            init.gridOptions == this->gridOptions);
-
-  if (allValuesIdenticalToInitialValues)
-    return;
-
-  YUViewDomElement newChild = root.ownerDocument().createElement(QString("statType%1").arg(typeID));
-  newChild.appendChild(root.ownerDocument().createTextNode(QString::fromStdString(typeName)));
-
-  // Append only the parameters that changed
-  if (init.render != render)
-    newChild.setAttribute("render", render);
-  if (init.alphaFactor != alphaFactor)
-    newChild.setAttribute("alphaFactor", alphaFactor);
-
-  addModifiedValuesToElement(newChild, this->valueDataOptions, this->init.valueDataOptions);
-  addModifiedValuesToElement(newChild, this->vectorDataOptions, this->init.vectorDataOptions);
-  addModifiedValuesToElement(newChild, this->gridOptions, this->init.gridOptions);
-
-  root.appendChild(newChild);
-}
-
-void StatisticsType::tryToLoadFromPlaylist(const YUViewDomElement &root)
-{
-  const auto [name, attributes] =
-      root.findChildValueWithAttributes(QString("statType%1").arg(this->typeID));
-
-  if (name.toStdString() != this->typeName)
-    // The name of this type with the right ID and the name in the playlist don't match?...
-    return;
-
-  // Parse and set all the attributes that are in the playlist
-  for (const auto [name, value] : attributes)
-  {
-    if (name == "render")
-      this->render = (value != "0");
-    else if (name == "alphaFactor")
-      this->alphaFactor = value.toInt();
-    else if (name == "renderValueData" || name == "scaleValueToBlockSize" ||
-             name == "colorMapperType")
-    {
-      if (!this->valueDataOptions)
-        this->valueDataOptions = ValueDataOptions();
-
-      if (name == "renderValueData")
-        this->valueDataOptions->render = (value != "0");
-      else if (name == "scaleValueToBlockSize")
-        this->valueDataOptions->scaleToBlockSize = (value != "0");
-      else if (name == "colorMapperType")
-        this->valueDataOptions->colorMapper.loadPlaylist(attributes);
-    }
-    else if (name == "renderVectorData" || name == "renderVectorDataValues" ||
-             name == "scaleVectorToZoom" || name == "vectorStyle" || name == "vectorScale" ||
-             name == "mapVectorToColor" || name == "renderarrowHead")
-    {
-      if (!this->vectorDataOptions)
-        this->vectorDataOptions = VectorDataOptions();
-
-      if (name == "renderVectorData")
-        this->vectorDataOptions->render = (value != "0");
-      else if (name == "renderVectorDataValues")
-        this->vectorDataOptions->renderDataValues = (value != "0");
-      else if (name == "scaleVectorToZoom")
-        this->vectorDataOptions->scaleToZoom = (value != "0");
-      else if (name == "vectorStyle")
-        this->vectorDataOptions->style = convertStringToPen(value);
-      else if (name == "vectorScale")
-        this->vectorDataOptions->scale = value.toInt();
-      else if (name == "mapVectorToColor")
-        this->vectorDataOptions->mapToColor = (value != "0");
-      else if (name == "renderarrowHead")
-      {
-        const auto idx = value.toInt();
-        if (idx >= 0 && unsigned(idx) < AllArrowHeads.size())
-          this->vectorDataOptions->arrowHead = AllArrowHeads.at(idx);
-      }
-    }
-    else if (name == "renderGrid")
-      this->gridOptions.render = (value != "0");
-    else if (name == "gridPen")
-      this->gridOptions.style = convertStringToPen(value);
-    else if (name == "scaleGridToZoom")
-      this->gridOptions.scaleToZoom = (value != "0");
-  }
-
-  this->saveInitialState();
 }
 
 std::string StatisticsType::getValueText(const int val) const
